Moved config method to separate Settings class
Start of settings cli options. So far can display settings file
